Deleting files

+++ controllers/admin.js
@@ -1,7 +1,8 @@
+const fileHelper = require('../util/file');
 const Product = require('../models/product');
-const product = require('../models/product');
 const { validationResult } = require('express-validator/check');
 const mongoose = require('mongoose');
+const product = require('../models/product');
 exports.getAddProduct = (req, res, next) => {
 
     res.render(
@@ -149,6 +150,7 @@ exports.postEditProducts = (req, res, next) => {
         product.title = updatetTitle;
         product.price = updatetPrice;
         if (image) {
+            fileHelper.deleteFile(product.imageUrl);
             product.imageUrl = image.path;
         }
         product.description = updatetDesc;
@@ -185,7 +187,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found!'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(() => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
